+++ consensus/src/runway/mod.rs
@@ -26,12 +26,10 @@ use std::{
     time::Duration,
 };
 
-mod backup;
 mod collection;
 mod packer;
 
-use crate::runway::backup::{BackupLoader, BackupSaver, LoadedData};
-pub use backup::BackupItem;
+use crate::backup::{BackupLoader, BackupSaver, LoadedData};
 #[cfg(feature = "initial_unit_collection")]
 use collection::{Collection, IO as CollectionIO};
 pub use collection::{NewestUnitResponse, Salt};
