+++ consensus/src/testing/crash_recovery.rs
@@ -1,5 +1,5 @@
 use crate::{
-    runway::BackupItem,
+    backup::BackupItem,
     testing::{init_log, spawn_honest_member, HonestMember, Network, ReconnectSender},
     units::UnitCoord,
     NodeCount, NodeIndex, SpawnHandle, TaskHandle,
